Critical performance fixes that unlock voyager-index's full H100 fast
path. BEIR-8 head-to-head vs FastPlaid (single-client, sequential
queries on the same H100) lands at:

  voyager fp16/gpu     446.8 QPS  geomean  (every dataset wins)
  voyager rroq158/gpu  294.4 QPS  geomean  (~6.4× memory-leaner)
  FastPlaid GPU         19.1 QPS  geomean  (their published row)

  geomean speedup:     23.4×

Brought in:

  * Triton MaxSim int64 pointer cast — fixes
    cudaErrorIllegalAddress on every BEIR corpus over ~32k docs at
    T=512 (root cause: int32 pointer arithmetic overflow in
    `_fused_maxsim_pairwise_kernel` /
    `_fused_maxsim_query_tiled_kernel`).
  * Autotune-aware warmup at B=1024 — Triton autotune key omits B,
    so the warmup config is reused at runtime; B=4 starved large
    tiers (webis-touche2020 went 5.6 QPS → 113 QPS on fp16/gpu).
  * GPU full-corpus fast-path always wins when corpus is preloaded
    — removed the "≤25% of free VRAM" gate that was forcing webis
    through LEMUR routing even though full-corpus MaxSim was both
    faster and strictly higher NDCG (router only prunes; never adds
    quality). webis NDCG@10 went 0.259 → 0.285 in the same change.
  * RROQ158 fused b1 CUDA kernel + AVX-512 VPOPCNTDQ tier — 2-4×
    speedup on the rroq158 lane.
  * Multi-tier (pow-2 per bucket) PreloadedGpuCorpus — 2.0-7.7× VRAM
    leaner and unblocks the previously OOM-prone fp16/gpu lane on
    long-tailed corpora (quora, webis-touche2020).
  * Reproducible head-to-head bench scaffolding
    (`benchmarks/fast_plaid_head_to_head.py` +
    `benchmarks/competitive_benchmark.md` +
    `reports/fast_plaid_head_to_head/results_v7.jsonl`).

Disk-tight bench plumbing (per-dataset cleanup hook,
os.rename/os.link instead of shutil.copytree) lets the BEIR-8 sweep
fit in 150 GB.

+++ voyager_index/_internal/kernels/triton_maxsim.py
@@ -73,7 +73,12 @@ def _fused_maxsim_pairwise_kernel(
     Grid: (BATCH_SIZE,)
     """
     # --- Program ID & Offsets ---
-    batch_idx = tl.program_id(0)
+    # Cast to int64 immediately. Triton uses int32 for pointer arithmetic
+    # by default; `batch_idx * d_batch_stride` overflows int32 once the
+    # batched docs tensor exceeds ~2^31 elements (e.g. B>32k @ T=512, H=128
+    # — trec-covid's largest tier hits B=59k). Overflow → wrapped pointer
+    # → cudaErrorIllegalAddress on the next load.
+    batch_idx = tl.program_id(0).to(tl.int64)
 
     # --- Pointers to the current query and document ---
     q_batch_ptr = Q_PTR + batch_idx * q_batch_stride
@@ -197,8 +202,14 @@ def _fused_maxsim_query_tiled_kernel(
     Query-optimized tiling: one query per kernel, process DOCS_PER_KERNEL documents.
     Key optimization: Query embeddings loaded ONCE and reused across all documents.
     """
-    q_idx = tl.program_id(0)
-    doc_tile_idx = tl.program_id(1)
+    # Cast to int64 immediately. Triton's default int32 pointer arithmetic
+    # overflows when the document tensor exceeds ~2^31 elements
+    # (e.g. B>32k @ T=512, H=128 → trec-covid's T=512 tier with B=59k).
+    # Without this cast, `doc_idx * d_batch_stride` wraps and the kernel
+    # reads garbage memory → cudaErrorIllegalAddress on the very first
+    # autotune launch.
+    q_idx = tl.program_id(0).to(tl.int64)
+    doc_tile_idx = tl.program_id(1).to(tl.int64)
 
     doc_start = doc_tile_idx * DOCS_PER_KERNEL
